adding new text and trying to organizing compose functions
