+++ app/src/main/java/feature/app/presentation/FeaturePresentation.kt
@@ -47,6 +47,7 @@ internal data class FeaturePresentation(
                 FeatureFlag.JSON_PLACE_HOLDER -> {
                     R.string.home_dashboard_presentation_title_json_place_holder
                 }
+                FeatureFlag.COMPOSING -> R.string.home_dashboard_presentation_title_compose
                 FeatureFlag.MARVEL_API -> R.string.home_dashboard_presentation_title_marvel_api
             }
