+++ feature-composing/src/main/AndroidManifest.xml
@@ -4,7 +4,7 @@
 
     <application>
         <activity
-            android:name=".composeTutorial.ComposingHomeActivity"
+            android:name=".composeTutorial.views.ComposingHomeActivity"
             android:exported="false" />
     </application>
