+++ feature-composing/src/main/java/feature/composing/composeTutorial/model/Message.kt
@@ -1,3 +1,3 @@
-package feature.composing.composeTutorial
+package feature.composing.composeTutorial.model
 
 data class Message(val author: String, val body: String)
