+++ feature-composing/src/main/java/feature/composing/composeTutorial/views/ComposingHomeActivity.kt
@@ -1,4 +1,4 @@
-package feature.composing.composeTutorial
+package feature.composing.composeTutorial.views
 
 import android.os.Bundle
 import androidx.activity.compose.setContent
@@ -33,6 +33,8 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import feature.composing.R
+import feature.composing.composeTutorial.SampleData
+import feature.composing.composeTutorial.composefunctions.Conversation
 
 class ComposingHomeActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -68,58 +70,5 @@ fun PreviewConversation() {
     }
 }
 
-@Composable
-fun Conversation(message: List<Message>) {
-    LazyColumn {
-        items(message) { message ->
-            MessageCard(msg = message)
-        }
-    }
-}
 
-@Composable
-fun MessageCard(msg: Message) {
-    Row(modifier = Modifier.padding(all = 8.dp)) {
-        Image(
-            painter = painterResource(id = R.drawable.profile_picture),
-            contentDescription = "Imagem de exemplo",
-            modifier = Modifier
-                .size(40.dp)
-                .clip(CircleShape)
-                .border(1.5.dp, MaterialTheme.colors.secondary, CircleShape)
-        )
 
-        Spacer(modifier = Modifier.width(8.dp))
-
-        var isExpanded by remember { mutableStateOf(false) }
-        val surfaceColor: Color by animateColorAsState(
-            if (isExpanded) {
-                MaterialTheme.colors.primary
-            } else {
-                MaterialTheme.colors.secondaryVariant
-            }
-        )
-
-        Column(modifier = Modifier.clickable { isExpanded = !isExpanded }) {
-            Text(
-                text = msg.author,
-                color = MaterialTheme.colors.secondaryVariant,
-                style = MaterialTheme.typography.subtitle2
-            )
-            Spacer(modifier = Modifier.height(4.dp))
-            Surface(
-                shape = MaterialTheme.shapes.medium,
-                elevation = 1.dp,
-                color = surfaceColor,
-                modifier = Modifier.animateContentSize().padding(1.dp)
-            ) {
-                Text(
-                    text = msg.body,
-                    modifier = Modifier.padding(all = 4.dp),
-                    maxLines = if (isExpanded) Int.MAX_VALUE else 1,
-                    style = MaterialTheme.typography.body2
-                )
-            }
-        }
-    }
-}
